requirements.txt
Django==1.7
django-bower==5.0.1
django-pipeline==1.3.25
djangorestframework==2.4.2
futures==2.1.6
gnureadline==6.3.3
ipython==2.2.0
model-mommy==1.2.2
six==1.7.3
wsgiref==0.1.2
